Fix dumpJson freeing its input
Fixes the segfault reported in #1 with ValueJson. The `dumpJson` call was
unexpectedly freeing its input, so `Next()` would try to free it again, leading
to crashes. Calling `dumpJson` again with the same input would also crash.

Adds a `jv_copy` to `dumpJson` so that the refcount is maintained after the
call, which was the expected behavior before. Also added some tests for the
previous crash in `ValueJson` and duplicate calls to `dumpJson`.

Thanks to @wordijp for reporting and investigating the issue in #1.